splash screen and log out

+++ Projekt/projekt/app/src/main/java/com/example/a3track/model/User.kt
@@ -3,7 +3,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class User(
-    var ID: Long,
+    var ID: Int,
     var last_name: String,
     var first_name: String,
     var email: String,
